'fix 500 server error on too long input and displaying wrong value on to big input'

+++ src/main/java/com/codecool/shop/dao/implementation/db/ProductDaoDB.java
@@ -29,7 +29,7 @@ public class ProductDaoDB implements ProductDao {
             ps.setInt(2, product.getSupplier().getId());
             ps.setString(3, product.getName());
             ps.setString(4, product.getDescription());
-            ps.setFloat(5, product.getDefaultPrice());
+            ps.setDouble(5, product.getDefaultPrice());
             ps.setString(6, product.getDefaultCurrency().toString());
 
             ResultSet rs = ps.executeQuery();
@@ -64,7 +64,7 @@ public class ProductDaoDB implements ProductDao {
         BaseDao<Supplier> sd = new SupplierDaoDB();
         Product product = new Product(
                 rs.getString("name"),
-                rs.getFloat("def_price"),
+                rs.getDouble("def_price"),
                 rs.getString("def_currency"),
                 rs.getString("description"),
                 pcd.find(rs.getInt("category_id")),
